move validator files to correct package (domain) since it handles business logic. did wrong before
