+++ app/src/main/java/com/example/tasky/common/domain/util/TextValidatorExtension.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.common.data
+package com.example.tasky.common.domain.util
 
 
 fun String.isValidPassword(): Boolean {
