+++ app/src/main/java/com/example/tasky/feature_login/presentation/LoginScreen.kt
@@ -145,6 +145,8 @@ fun LoginScreenContent(
                     loginViewModel.onShowErrorDialog(message)
                 }
             }
+
+            null -> println("no action")
         }
 
         if (uiState.showErrorDialog) {
