+++ app/src/main/java/com/example/tasky/feature_login/presentation/LoginViewModel.kt
@@ -3,10 +3,10 @@ package com.example.tasky.feature_login.presentation
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.tasky.common.data.EmailPatternValidatorImpl
-import com.example.tasky.common.data.isValidName
-import com.example.tasky.common.data.isValidPassword
 import com.example.tasky.common.domain.Result
+import com.example.tasky.common.domain.util.EmailPatternValidatorImpl
+import com.example.tasky.common.domain.util.isValidName
+import com.example.tasky.common.domain.util.isValidPassword
 import com.example.tasky.feature_login.domain.repository.UserRemoteRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
